Move Containerd specific code to its own package
